+++ .github/workflows/periodic.yml
@@ -14,9 +14,6 @@ concurrency:
 
 jobs:
   linux-bionic-cuda11_6-py3-gcc7-slow-gradcheck-build:
-    # Disable because slow-gradcheck tests take > 4 hrs and time out.
-    # TODO(sdym@meta.com): investigate re-enabling slow-gradcheck tests.
-    if: false
     name: linux-bionic-cuda11.6-py3-gcc7-slow-gradcheck
     uses: ./.github/workflows/_linux-build.yml
     with:
@@ -29,9 +26,6 @@ jobs:
         ]}
 
   linux-bionic-cuda11_6-py3-gcc7-slow-gradcheck-test:
-    # Disable because slow-gradcheck tests take > 4 hrs and time out.
-    # TODO(sdym@meta.com): investigate re-enabling slow-gradcheck tests.
-    if: false
     name: linux-bionic-cuda11.6-py3-gcc7-slow-gradcheck
     uses: ./.github/workflows/_linux-test.yml
     needs: linux-bionic-cuda11_6-py3-gcc7-slow-gradcheck-build
