+++ test/dynamo/test_recompile_ux.py
@@ -8,8 +8,10 @@ import torch._dynamo
 import torch._dynamo.config
 import torch._dynamo.test_case
 import torch._dynamo.testing
+from torch.testing._internal.common_utils import skipIfSlowGradcheckEnv
 
 
+@skipIfSlowGradcheckEnv
 class RecompileUxTests(torch._dynamo.test_case.TestCase):
     # TODO(whc) dynamo actualy recompiles one more time than the cache limit
     cache_limit = 1
