+++ test/inductor/test_smoke.py
@@ -5,6 +5,7 @@ import unittest
 import torch
 import torch._dynamo as torchdynamo
 import torch._inductor.config as torchinductor_config
+from torch.testing._internal.common_utils import skipIfSlowGradcheckEnv
 
 torchdynamo.config.log_level = logging.INFO
 torchdynamo.config.verbose = True
@@ -23,6 +24,7 @@ class MLP(torch.nn.Module):
         return x
 
 
+@skipIfSlowGradcheckEnv
 class SmokeTest(unittest.TestCase):
     def test_mlp(self):
         mlp = torchdynamo.optimize("inductor")(MLP().cuda())
